Off by 1, early large num termination
Fixed a couple off by one bugs. Also wrote an early termination for very large numbers (mirrors terminating less than 0 mantissa bits).

The smallest integer which cannot be exactly represented by FP32 is 16777217. This matches what Wikipedia says.

Whenever a number cannot be represented, the applet will round down. This is not the most accurate metric, but it is fast and more or less correct.